+++ AccountSystemTranslator/src/main/java/za/ac/nwu/as/translator/AccountHolderTranslator.java
@@ -11,7 +11,7 @@ public interface AccountHolderTranslator {
 
     List<AccountHolder> getAllAccountHolders();
 
-    AccountHolder getAccountHolderByPk(Long memberId);
+    AccountHolder getAccountHolderByPk(int memberId);
 
     AccountHolderDto create(AccountHolderDto accountHolder);
